+++ packages/expo-router/src/imperative-api.ts
@@ -2,12 +2,12 @@ import { store } from './global-state/router-store';
 import { ExpoRouter } from '../types/expo-router';
 
 export const router: ExpoRouter.Router = {
-  navigate: (href) => store.navigate(href),
-  push: (href) => store.push(href),
+  navigate: (href, options) => store.navigate(href, options),
+  push: (href, options) => store.push(href, options),
   dismiss: (count) => store.dismiss(count),
   dismissAll: () => store.dismissAll(),
   canDismiss: () => store.canDismiss(),
-  replace: (href) => store.replace(href),
+  replace: (href, options) => store.replace(href, options),
   back: () => store.goBack(),
   canGoBack: () => store.canGoBack(),
   setParams: (params) => store.setParams(params),